+++ Package.swift
@@ -5,15 +5,15 @@ import PackageDescription
 
 let package = Package(
     name: "ALTENLogger",
-    platforms: [.iOS(.v11), .macOS(.v10_12), .tvOS(.v12), .watchOS(.v7)],
+    platforms: [.iOS(.v14), .macOS(.v11), .tvOS(.v14), .watchOS(.v7)],
     products: [
         .library(
             name: "ALTENLoggerFirebase",
             targets: ["ALTENLoggerFirebase"])
     ],
     dependencies: [
-        .package(url: "https://github.com/SDOSLabs/ALTENLoggerCore.git", .upToNextMajor(from: "1.0.0")),
-        .package(url: "https://github.com/firebase/firebase-ios-sdk.git", .upToNextMajor(from: "10.1.0"))
+        .package(url: "https://github.com/SDOSLabs/ALTENLoggerCore.git", .upToNextMajor(from: "1.1.0")),
+        .package(url: "https://github.com/firebase/firebase-ios-sdk.git", .upToNextMajor(from: "10.15.0"))
     ],
     targets: [
         .target(