+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [2.0.0](https://github.com/SDOSLabs/ALTENLoggerFirebase/tree/2.0.0)
+
+- Change log format in console
+- Add new parameter `hideConsoleLog`. Now you can hide console logs (it doesn't affect Firebase logging)
+- Upgrade Firebase dependency to 10.15.0
+
 ## [1.0.0](https://github.com/SDOSLabs/ALTENLoggerFirebase/tree/1.0.0)
 
 - Create first stable version